-Have created package for entities and Post class is moved from school package to package for entities. -created post endpoint. -Commented out variables for local date due to that we cannot solve the date issue in Springboot.
