+++ .idea/csv-plugin.xml
@@ -3,6 +3,13 @@
   <component name="CsvFileAttributes">
     <option name="attributeMap">
       <map>
+        <entry key="\src\main\java\com\grupparbete\PostControllers\PostsController.java">
+          <value>
+            <Attribute>
+              <option name="separator" value=";" />
+            </Attribute>
+          </value>
+        </entry>
         <entry key="\src\main\java\com\grupparbete\school\SurferApplication.java">
           <value>
             <Attribute>
