+++ src/main/java/com/grupparbete/PostControllers/PostsController.java
@@ -1,10 +1,8 @@
 package com.grupparbete.PostControllers;
 
-import com.grupparbete.school.Post;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
+import com.grupparbete.entities.Post;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -12,14 +10,23 @@ import java.util.List;
 import java.util.Map;
 
 @RestController
-@RequestMapping("api/v1/post")
+@RequestMapping("api/v1/posts")
 public class PostsController {
 
-Map<Integer, Post> post=new HashMap<Integer,Post>();
+Map<Integer, Post> posts=new HashMap<Integer,Post>();
         @GetMapping("/")
         public Collection<Post> getAll() {
-            return post.values();
-
+            return posts.values();
+            
+    }
+    @GetMapping ("/{id}")
+    public Post getById (@PathVariable long id) {
+            return posts.get(id);
+    }
+    @PostMapping("/")
+    public Post addNewPost(@RequestBody Post post){
+            posts.put((int) post.getId(),post);
+            return post;
 
     }
 }
